Create system_utils replacement for mkstemp, remove related aducpal apis.

+++ src/libaducpal/src/time.c
@@ -65,7 +65,7 @@ int ADUCPAL_nanosleep(const struct timespec* rqtp, struct timespec* rmtp)
     }
 
     // Add in number of seconds converted to ms.
-    Sleep(ms + (rqtp->tv_sec * 1000));
+    Sleep(ms + ((DWORD)(rqtp->tv_sec) * 1000));
 
     return 0;
 }
